DriverPack Downloader version 2.23
The HP DriverPack downloader script is an updated and modified version of GaryTown.com script. It sports a GUI interface and is extended to automatically create the needed Task Sequences - as used by the original script - and to create the Driverpack as Software Packages used by SCCM as well as added the appropriate TS call steps. The script is described in detail at https://developers.hp.com/hp-client-management/blog/scripting-download-and-use-hp-driverpacks-sccmsecm, while the original script is described in https://garytown.com/hp-driver-packs-download-cm-import-via-powershell